# api/api/openapi.yaml
openapi: 3.0.3
info:
  title: UpcycleConnect API
  version: 1.0.0
  description: API REST pour UpcycleConnect (admin, pro, particulier)
servers:
  - url: http://localhost:8080
paths:
...
  /events:
    get:
      summary: Lister les sessions / événements
      responses:
        '200': { description: OK }
    post:
      summary: Créer une session (staff/salarié EMP)
      description: Le salarié (role 5) crée toujours en statut en_attente avec id_createur.
      responses:
        '201': { description: Créé }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
  /events/{id}:
    put:
      summary: Mettre à jour une session (admin/staff ou salarié sur ses en_attente)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Mis à jour }
        '403': { description: Forbidden }
    delete:
      summary: Supprimer une session (admin/staff ou salarié sur ses en_attente)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Supprimé }
        '403': { description: Forbidden }
  /me/events:
    get:
      summary: Mes événements (salarié / créateur)
      description: Sessions où id_createur = utilisateur connecté.
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via POST /login
...